optimize case import component
